e1: change ISO EP size to 512
256 is too small, as one packet is scheduled (max!) per 1ms frame,
resulting in 256kByte/s, which is just not enough for an E1 signal
that runs off a clock that might be faster than the host PC.  The SAM4S
doesn't seem to support odd wMaxPacketSize (I tried 388 but failed) and
hence we set it to 512.  This means we will only transmit data in an
average ot 50% of the frames where we get scheduled to transfer.